refine

+++ oneflow/user/kernels/variance_kernel_util.cpp
@@ -24,8 +24,8 @@ template<typename T>
 struct VarFunctor<DeviceType::kCPU, T> final {
   void operator()(ep::Stream* stream, const T* in_ptr, T* out_ptr, T* tmp_buffer_ptr,
                   VarParam var_param) {
-    for (int i = 0; i < var_param.parallel_num; i++) {
-      int input_offset = LinearIndex2Offset(i, var_param.dim_size_in_caxis,
+    for (size_t i = 0; i < var_param.parallel_num; i++) {
+      size_t input_offset = LinearIndex2Offset(i, var_param.dim_size_in_caxis,
                                             var_param.stride_in_caxis, var_param.caxis_size);
       ComputeVarUsingWelford(&in_ptr[input_offset], &out_ptr[i], var_param);
     }
